requirements.txt
# AWS SDK
boto3==1.34.34
awscli==1.32.34

# Data Processing
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2

# ML & Feature Engineering
scikit-learn==1.4.0
xgboost==2.0.3
sagemaker==2.199.0

# Data Quality
great-expectations==0.18.8

# Testing
pytest==7.4.4
moto==5.0.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Development
black==24.1.1
ipython==8.20.0
jupyter==1.0.0

# Database
psycopg2-binary==2.9.9
joblib==1.3.2
